Slice Tracking with Dictation/requirements.txt
google-cloud-speech
sounddevice
PyAudio
numpy
pandas
plotly
wave
# API client (lightweight)
requests
# Image processing (for extracting slice images)
Pillow